types: allow chug logs to be parsed and add LogEntryDecoder
When the '-json' flag is provided, chug writes the combined/parsed logs
to JSON, but this occurs after the Session, Error and Trace fields have
been parsed and removed from the Data field.  This commit adds them to
the combinied log format struct so that the output of 'chug -json ...'
is correctly parsed.